Dramatically enrich signing to support notarization
Specifically, a bunch more files need to be signed, now,
and the all need to be signed with the hardened runtime
requirement, an apple-approved timestamp, and a list
of entitlements.

There's still a WIP aspect to this, because it doesn't
use the xml/plist framework; my suspicion is that the
representation of true as <true></true> breaks the
entitlement engine.