Revert "simplify ".dmg" creation, avoiding a step that mounts the image"
This reverts commit 368e1a48c8a91d028c15b05cbbc7dbdb9a4e2bf8.

The reason to mount the image after it's create is to create an alias
that works for the background picture. Although the alias is just a
pile of bytes, the format of an alias is not documented. Also, the
alias wants to refer to the image file's inode value, which is only
determined after the filesystem image is created.