improve log file naming
Due to the way responsibility is distributed between clients and the
server, the (default) log file wants to be based on the download-page
description, and not the file name. But adjust characters in that
description to make it a better file name. Also, make `#:log-file`
work right with snapshot pages.