installer-dmg: recur to sign in directories of "lib"
The new "ossl-modules" directory installed by packages with OpenSSL 3
has a shared library that needs to be signed --- at least in the case
of x86_64 MacOS, where the package does not have the shared library
signed already, and notarization insists that it's signed.

Also, sign loadable "modules" as well as libraries, since
"legacy.dylib" is a module.